+++ lalr.c
@@ -112,7 +112,7 @@ static void set_maxrhs(void)
 
 static void initialize_LA(void)
 {
-  register size_t i, j, k;
+  register int i, j, k;
   register reductions *rp;
 
   lookaheads = NEW2(nstates + 1, Yshort);
@@ -152,9 +152,9 @@ static void initialize_LA(void)
 static void set_goto_map(void)
 {
   register shifts *sp;
-  register size_t i;
+  register int i;
   register int symbol;
-  register size_t k;
+  register int k;
   register Yshort *temp_map;
   register int state2;
   register int state1;
@@ -249,7 +249,7 @@ static void traverse(int i)
   register unsigned *fp1;
   register unsigned *fp2;
   register unsigned *fp3;
-  register size_t j;
+  register int j;
   register Yshort *rp;
 
   int height;
@@ -302,7 +302,7 @@ static void traverse(int i)
 
 static void digraph(Yshort** relation)
 {
-  register size_t i;
+  register int i;
 
   infinity = ngotos + 2;
   INDEX = NEW2(ngotos + 1, Yshort);
@@ -327,8 +327,8 @@ static void digraph(Yshort** relation)
 
 static void initialize_F(void)
 {
-  register size_t i;
-  register size_t j;
+  register int i;
+  register int j;
   register int k;
   register shifts *sp;
   register Yshort *edge;
@@ -402,7 +402,7 @@ static void initialize_F(void)
 
 static void add_lookback_edge(int stateno, int ruleno, int gotono)
 {
-    register size_t i, k;
+    register int i, k;
     register int found;
     register shorts *sp;
 
@@ -431,8 +431,8 @@ static Yshort** transpose(Yshort** R, int n)
   register Yshort **temp_R;
   register Yshort *nedges;
   register Yshort *sp;
-  register size_t i;
-  register size_t k;
+  register int i;
+  register int k;
 
   nedges = NEW2(n, Yshort);
 
@@ -481,8 +481,8 @@ static Yshort** transpose(Yshort** R, int n)
 
 static void build_relations(void)
 {
-  register size_t i;
-  register size_t j;
+  register int i;
+  register int j;
   register int k;
   register Yshort *rulep;
   register Yshort *rp;
@@ -582,7 +582,7 @@ static void compute_FOLLOWS(void)
 
 static void compute_lookaheads(void)
 {
-  register size_t i, n;
+  register int i, n;
   register unsigned *fp1, *fp2, *fp3;
   register shorts *sp, *next;
   register unsigned *rowp;
