+++ symtab.c
@@ -83,7 +83,7 @@ bucket* lookup(char const * name)
 
 void create_symbol_table()
 {
-    register size_t i;
+    register int i;
     register bucket *bp;
 
     symbol_table = (bucket **) MALLOC(TABLE_SIZE*sizeof(bucket *));
